Update main.c
had a breakthrough. To work on the decryption properly and not just overwrite the data, I needed to copy the alphaCopy array and work the decrypt function through those two arrays (the second one called encryptionKey) rather than through the two original arrays. this should hopefully make things a little easier in the future